requirements.txt
flask==3.0.2
peewee==3.17.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0